[MT-PGA] refactor and convert to diagram feature
