+++ core/base/mergeTreePrincipalGeodesics/CMakeLists.txt
@@ -1,8 +1,10 @@
 ttk_add_base_library(mergeTreePrincipalGeodesics
   SOURCES
+    MergeTreeAxesAlgorithmBase.cpp
     MergeTreePrincipalGeodesicsBase.cpp
     MergeTreePrincipalGeodesics.cpp
   HEADERS
+    MergeTreeAxesAlgorithmBase.h
     MergeTreePrincipalGeodesicsBase.h
     MergeTreePrincipalGeodesics.h
     MergeTreePrincipalGeodesicsUtils.h
